added challenge 53 (WIP, works but a lil off spec)
The function to search for a link block just looks for a block in the
output that equals the final block from the expandable message, rather
than finding a "bridge" block. This is only a problem for very
repetitive messages (because for a janky function with small sizes for
message blocks and internal state, repetitive messages' H values can
easily fall into short cycles). If the first preimage has a decent
amount of variety in its message blocks, this isn't a problem.